Update get_individual_algebra_results.R
Was trying to download and install mxsem and was getting issues flagged with get_individual_algebra_results.R. Might be an issue on line 71—72 of the original file. Seems to be related to the use of `|>` (pipe) operator. Updated the line with conventional use of `gsub` and this seems to have resolved the installation issue I had. I manually updated the .r file and it seems to have worked out okay in the manual installation.